requirements.txt
nostr-sdk>=0.27.0
tweepy>=4.14.0
websockets>=12.0
requests>=2.31.0
python-dotenv>=1.0.0
